Disabled built-in gapless support in mpg123, since it doesnt work properly with seeking
This plugin doesn't seek by itself; the mpegaudioparse and mp3parse plugins
can do that. However, they do not handle gapless properly. This causes
gaps with mpg123 decoding, even when gapless was enabled in the mpg123 handle.

Disabling gapless in the decoder for now, until the mpegaudioparse element
has proper support for it.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>